fastapi_app/requirements.txt
flask
requests
gunicorn
fastapi
uvicorn
python-jose[cryptography]
Flask-JWT-Extended
